create test

+++ packets/server/reward_got.go
@@ -4,6 +4,6 @@ const REWARD_GOT PacketServerType = 315
 
 // REWARD_GOT(315) "II"
 type RewardGot struct {
-	UserID   int
+	UserID   uint64
 	RewardID int
 }
